requirements.txt
flask~=2.3.2
gunicorn
psycopg; sys_platform == "linux"
psycopg[binary]; sys_platform != "linux"